requirements.txt
paho-mqtt
systemd
pyyaml
